Move the interfaces to DataObject\ClassDefinition\Data\
