+++ models/DataObject/ClassDefinition/Data/Block.php
@@ -21,8 +21,6 @@ use Pimcore\Logger;
 use Pimcore\Model;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\ClassDefinition\Data;
-use Pimcore\Model\DataObject\CustomResourcePersistingInterface;
-use Pimcore\Model\DataObject\ResourcePersistenceAwareInterface;
 use Pimcore\Model\Element;
 use Pimcore\Tool\Serialize;
