+++ models/DataObject/ClassDefinition/Data/Classificationstore.php
@@ -19,7 +19,6 @@ namespace Pimcore\Model\DataObject\ClassDefinition\Data;
 use Pimcore\Model;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\ClassDefinition\Data;
-use Pimcore\Model\DataObject\CustomResourcePersistingInterface;
 use Pimcore\Model\Element;
 use Pimcore\Tool;
