+++ models/DataObject/ClassDefinition/Data/EncryptedField.php
@@ -21,7 +21,6 @@ use Defuse\Crypto\Key;
 use Pimcore\Logger;
 use Pimcore\Model;
 use Pimcore\Model\DataObject\ClassDefinition\Data;
-use Pimcore\Model\DataObject\ResourcePersistenceAwareInterface;
 
 /**
  * Class EncryptedField
