+++ models/DataObject/ClassDefinition/Data/Fieldcollections.php
@@ -20,7 +20,6 @@ use Pimcore\Logger;
 use Pimcore\Model;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\ClassDefinition\Data;
-use Pimcore\Model\DataObject\CustomResourcePersistingInterface;
 use Pimcore\Model\Webservice;
 use Pimcore\Tool\Cast;
