+++ models/DataObject/ClassDefinition/Data/Input.php
@@ -18,8 +18,6 @@ namespace Pimcore\Model\DataObject\ClassDefinition\Data;
 
 use Pimcore\Model;
 use Pimcore\Model\DataObject\ClassDefinition\Data;
-use Pimcore\Model\DataObject\ResourcePersistenceAwareInterface;
-use Pimcore\Model\DataObject\QueryResourcePersistenceAwareInterface;
 
 class Input extends Data implements ResourcePersistenceAwareInterface, QueryResourcePersistenceAwareInterface
 {
