+++ models/DataObject/ClassDefinition/Data/Password.php
@@ -19,8 +19,6 @@ namespace Pimcore\Model\DataObject\ClassDefinition\Data;
 use Pimcore\Model;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\ClassDefinition\Data;
-use Pimcore\Model\DataObject\ResourcePersistenceAwareInterface;
-use Pimcore\Model\DataObject\QueryResourcePersistenceAwareInterface;
 
 class Password extends Data implements ResourcePersistenceAwareInterface, QueryResourcePersistenceAwareInterface
 {
