+++ models/DataObject/ClassDefinition/Data/QueryResourcePersistenceAwareInterface.php
@@ -15,34 +15,25 @@
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-namespace Pimcore\Model\DataObject;
+namespace Pimcore\Model\DataObject\ClassDefinition\Data;
 
-interface ResourcePersistenceAwareInterface
+interface QueryResourcePersistenceAwareInterface
 {
     /**
-     * Returns the the data that should be stored in the resource
+     * Returns the data which should be stored in the query columns
      *
      * @param mixed $data
      * @param null|\Pimcore\Model\DataObject\AbstractObject $object
      * @param mixed $params
      *
      * @return mixed
-     */
-    public function getDataForResource($data, $object = null, $params = []);
-
-    /**
-     * Convert the saved data in the resource to the internal eg. Image-Id to Asset\Image object, this is the inverted getDataForResource()
      *
-     * @param mixed $data
-     * @param null|\Pimcore\Model\DataObject\AbstractObject $object
-     * @param mixed $params
-     *
-     * @return mixed
+     * abstract public function getDataForQueryResource($data);
      */
-    public function getDataFromResource($data, $object = null, $params = []);
+    public function getDataForQueryResource($data, $object = null, $params = []);
 
     /**
      * @return string|array
      */
-    public function getColumnType();
+    public function getQueryColumnType();
 }
