+++ models/DataObject/ClassDefinition/Data/Relations/AbstractRelations.php
@@ -21,7 +21,7 @@ use Pimcore\Logger;
 use Pimcore\Model;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\ClassDefinition\Data;
-use Pimcore\Model\DataObject\CustomResourcePersistingInterface;
+use Pimcore\Model\DataObject\ClassDefinition\Data\CustomResourcePersistingInterface;
 use Pimcore\Model\Element;
 
 abstract class AbstractRelations extends Data implements CustomResourcePersistingInterface
