+++ models/DataObject/ClassDefinition/Data/Table.php
@@ -19,8 +19,6 @@ namespace Pimcore\Model\DataObject\ClassDefinition\Data;
 use Pimcore\Model;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\ClassDefinition\Data;
-use Pimcore\Model\DataObject\ResourcePersistenceAwareInterface;
-use Pimcore\Model\DataObject\QueryResourcePersistenceAwareInterface;
 use Pimcore\Tool\Serialize;
 
 class Table extends Data implements ResourcePersistenceAwareInterface, QueryResourcePersistenceAwareInterface
